Data-Modeling-Projects/Relational-DataBase/requirements.txt
pandas==1.5.3
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
